*: move `session.NewStore` to `store.New` to achieve semantic accuracy
The API `session.NewStore` should be placed in package `store` and
rename `NewStore` to `New` for achieving more semantic accuracy.
